Replace explicit i8 with libc::c_char, to allow it to compile on ARM.
Despite the docs saying that libc::c_char is equivalent to i8, it would
appear that libc::c_char actually follows the idiosyncracies of the C
`char` type that it represents. That is, on some platforms it is
equivalent to `signed char` (i8) and on others it is `unsigned char`
(u8). By explicitly using libc::c_char in the `stoptr` and `ptrtos`
functions, we avoid mismatched type errors when compiling on ARM.